requirements.txt
streamlit 
groq 
python-dotenv
sentence-transformers
faiss-cpu
numpy
pypdf
accelerate